+++ financial-connections-example/src/main/java/com/stripe/android/financialconnections/example/settings/PlaygroundSettings.kt
@@ -177,7 +177,6 @@ internal data class PlaygroundSettings(
             ConfirmIntentSetting(),
             NativeSetting(),
             PermissionsSetting(),
-            DynamicAppearanceSetting(),
             EmailSetting(),
             CustomerIdSetting(),
             StripeAccountIdSetting().takeIf { BuildConfig.TEST_ENVIRONMENT != "edge" },