+++ paymentsheet/src/main/java/com/stripe/android/customersheet/data/CustomerAdapterDataSource.kt
@@ -48,7 +48,7 @@ internal class CustomerAdapterDataSource @Inject constructor(
             }
 
             val savedSelectionResult = async {
-                retrieveSavedSelection().toResult()
+                retrieveSavedSelection(customerSessionElementsSession = null).toResult()
             }
 
             val elementsSession = elementsSessionResult.await().getOrThrow()
@@ -91,7 +91,9 @@ internal class CustomerAdapterDataSource @Inject constructor(
         customerAdapter.detachPaymentMethod(paymentMethodId)
     }
 
-    override suspend fun retrieveSavedSelection() = runCatchingAdapterTask {
+    override suspend fun retrieveSavedSelection(
+        customerSessionElementsSession: CustomerSessionElementsSession?
+    ) = runCatchingAdapterTask {
         customerAdapter.retrieveSelectedPaymentOption().map { result ->
             result?.toSavedSelection()
         }