+++ paymentsheet/src/test/java/com/stripe/android/link/ui/wallet/WalletScreenScreenshotTest.kt
@@ -188,7 +188,6 @@ internal class WalletScreenScreenshotTest {
                     onPayAnotherWayClicked = {},
                     onRemoveClicked = {},
                     onSetDefaultClicked = {},
-                    onEditPaymentMethodClicked = {},
                     showBottomSheetContent = {},
                     hideBottomSheetContent = {},
                     onAddNewPaymentMethodClicked = {},