+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/verticalmode/PaymentMethodEmbeddedLayoutUIScreenshotTest.kt
@@ -1,7 +1,10 @@
 package com.stripe.android.paymentsheet.verticalmode
 
 import android.graphics.Color
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.toArgb
 import com.stripe.android.model.PaymentMethodFixtures
 import com.stripe.android.paymentelement.ExperimentalEmbeddedPaymentElementApi
@@ -137,6 +140,7 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
         displayableSavedPaymentMethod: DisplayableSavedPaymentMethod? = savedPaymentMethod,
         newPaymentMethods: List<DisplayablePaymentMethod> = paymentMethods
     ) {
+        val scrollState = rememberScrollState()
         PaymentMethodEmbeddedLayoutUI(
             paymentMethods = newPaymentMethods,
             displayedSavedPaymentMethod = displayableSavedPaymentMethod,
@@ -148,7 +152,8 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
             onSelectSavedPaymentMethod = {},
             onManageOneSavedPaymentMethod = {},
             imageLoader = mock(),
-            rowStyle = rowStyle
+            rowStyle = rowStyle,
+            modifier = Modifier.verticalScroll(scrollState),
         )
     }
 
@@ -164,7 +169,8 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
         bottomSeparatorEnabled: Boolean? = null,
         selectedColor: Int? = null,
         unselectedColor: Int? = null,
-        additionalInsetsDp: Float? = null,
+        additionalVerticalInsetsDp: Float? = null,
+        horizontalInsetsDp: Float? = null,
         checkmarkColor: Int? = null,
         checkmarkInsetDp: Float? = null,
         spacingDp: Float? = null
@@ -179,7 +185,9 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 bottomSeparatorEnabled = bottomSeparatorEnabled ?: StripeThemeDefaults.flat.bottomSeparatorEnabled,
                 selectedColor = selectedColor ?: StripeThemeDefaults.colorsLight.materialColors.primary.toArgb(),
                 unselectedColor = unselectedColor ?: StripeThemeDefaults.colorsLight.componentBorder.toArgb(),
-                additionalInsetsDp = additionalInsetsDp ?: StripeThemeDefaults.embeddedCommon.additionalInsetsDp
+                additionalVerticalInsetsDp = additionalVerticalInsetsDp
+                    ?: StripeThemeDefaults.embeddedCommon.additionalVerticalInsetsDp,
+                horizontalInsetsDp = horizontalInsetsDp ?: StripeThemeDefaults.embeddedCommon.horizontalInsetsDp
             )
             Embedded.RowStyle.FlatWithCheckmark::class -> Embedded.RowStyle.FlatWithCheckmark(
                 separatorThicknessDp = separatorThicknessDp ?: StripeThemeDefaults.flat.separatorThickness,
@@ -190,11 +198,14 @@ class PaymentMethodEmbeddedLayoutUIScreenshotTest {
                 bottomSeparatorEnabled = bottomSeparatorEnabled ?: StripeThemeDefaults.flat.bottomSeparatorEnabled,
                 checkmarkColor = checkmarkColor ?: StripeThemeDefaults.colorsLight.materialColors.primary.toArgb(),
                 checkmarkInsetDp = checkmarkInsetDp ?: StripeThemeDefaults.embeddedCommon.checkmarkInsetDp,
-                additionalInsetsDp = additionalInsetsDp ?: StripeThemeDefaults.embeddedCommon.additionalInsetsDp
+                additionalVerticalInsetsDp = additionalVerticalInsetsDp
+                    ?: StripeThemeDefaults.embeddedCommon.additionalVerticalInsetsDp,
+                horizontalInsetsDp = horizontalInsetsDp ?: StripeThemeDefaults.embeddedCommon.horizontalInsetsDp
             )
             else -> Embedded.RowStyle.FloatingButton(
                 spacingDp = spacingDp ?: StripeThemeDefaults.floating.spacing,
-                additionalInsetsDp = additionalInsetsDp ?: StripeThemeDefaults.embeddedCommon.additionalInsetsDp
+                additionalInsetsDp = additionalVerticalInsetsDp
+                    ?: StripeThemeDefaults.embeddedCommon.additionalVerticalInsetsDp
             )
         }
     }